Moved response types into the same file.
